backend/requirements.txt
# FastAPI & ASGI
fastapi[all]>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Redis
redis>=5.0.0
hiredis>=2.3.0

# Kafka
aiokafka>=0.10.0
kafka-python>=2.0.2

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-keycloak>=3.9.0
cryptography>=42.0.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.3
ulid-py>=1.1.0

# OpenSearch
opensearch-py>=2.4.0

# Object Storage (MinIO/S3)
boto3>=1.34.0
minio>=7.2.0

# Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
